refactor: moved stuff into src
