--- Personer/function.json
@@ -1,18 +1,0 @@
-{
-  "scriptFile": "../dist/Personer/index.js",
-  "entryPoint": "handler",
-  "bindings": [
-    {
-      "authLevel": "anonymous",
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req",
-      "methods": ["post"]
-    },
-    {
-      "type": "http",
-      "direction": "out",
-      "name": "$return"
-    }
-  ]
-}
