+++ src/lib/decode-bearer-token.ts
